requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
pydantic>=2.5.2
requests>=2.31.0
python-dotenv>=1.0.0
astrapy>=0.7.4
langflow>=0.5.3